Hangman
This is a program that uses a dictionary to randomly choose a word for you to guess via the popular game hangman. Gives you the definition of the word regardless of if you get it or not. Also allows you to play PvP against a friend--doesn't give a definition for this, so you can put whatever you want in there. One of my biggest projects as of 4/13/2021